% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{DAMinMax}
\alias{DAMinMax}
\alias{TrajDAFindFirstMinimum}
\alias{TrajDAFindFirstMaximum}
\title{First direction autocorrelation minimum/maximum}
\usage{
TrajDAFindFirstMinimum(corr, windowSize = 10)

TrajDAFindFirstMaximum(corr, windowSize = 10)
}
\arguments{
\item{corr}{Direction autocorrelation of a trajectory.}

\item{windowSize}{Size of window used to define what constitutes a local
mimimum/maximum.}
}
\value{
Numeric vector with 2 values, \code{deltaS} and \code{C}, or else
  NULL if there is no local minimum/maximum.
}
\description{
Determines the coordinates of the first local minimum/maximum of \code{C} in
the direction autocorrelation function of a trajectory as returned by
\code{\link{TrajDirectionAutocorrelations}}. The end point is excluded from
consideration as a minimum, similarly the start point will not be returned as a
maximum.
}
\examples{
# Resample to fixed path length
resampled <- TrajRediscretize(trj, .005)
# Calculate direction autocorrelation for resampled trajectory
corr <- TrajDirectionAutocorrelations(resampled)
# Extract first local minimum from autocorrelation
minPt <- TrajDAFindFirstMinimum(corr, 50)

# Plot the autocorrelation function
plot(corr)
# Plot a red dot with a black outline at the first minimum
points(minPt["deltaS"], minPt["C"], pch = 16, col = "red", lwd = 2)
points(minPt["deltaS"], minPt["C"], col = "black", lwd = 2)

}
\seealso{
\code{\link{TrajDirectionAutocorrelations}}
}
