% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{points.Trajectory}
\alias{points.Trajectory}
\title{Add Trajectory points to a plot}
\usage{
\method{points}{Trajectory}(x, draw.start.pt = TRUE,
  turning.angles = NULL, ...)
}
\arguments{
\item{x}{An object of class "Trajectory", the trajectory to be plotted.}

\item{draw.start.pt}{If TRUE, draws a dot at the start point of the
trajectory.}

\item{turning.angles}{If \code{random} or \code{directed}, draws step turning
angles. \code{directed} assumes errors are relative to the first recorded
step angle. \code{random} assumes errors are relative to the previous step.}

\item{...}{Additional arguments are passed to \code{\link[graphics]{points}}.}
}
\description{
The \code{points} method for Trajectory objects.
}
