% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{Traj3DDistance}
\alias{Traj3DDistance}
\title{3D Trajectory distance}
\usage{
Traj3DDistance(trj3d, startIndex = 1, endIndex = nrow(trj3d))
}
\arguments{
\item{trj3d}{3-dimensional trajectory whose distance is to be calculated.}

\item{startIndex}{Index of the starting point.}

\item{endIndex}{Index of the ending point.}
}
\value{
Numeric distance from the start to the end of the trajectory.
}
\description{
Calculates the distance between the start and end of a 3-dimensional trajectory (or a
portion of a trajectory). Also called the diffusion distance, net distance,
displacement, or bee-line from start to finish.
}
\seealso{
\code{\link{Traj3DFromCoords}}
}
