% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{Traj3DLength}
\alias{Traj3DLength}
\title{3D trajectory length}
\usage{
Traj3DLength(trj3d, startIndex = 1, endIndex = nrow(trj3d))
}
\arguments{
\item{trj3d}{3D trajectory whose length is to be calculated.}

\item{startIndex}{Index of the starting point.}

\item{endIndex}{Index of the ending point.}
}
\value{
Numeric length of the trajectory.
}
\description{
Calculates the cumulative length of a 3-dimensional trajectory, or a portion
of a trajectory. The length is the total distance travelled along the trajectory.
}
\seealso{
\code{\link{Traj3DFromCoords}}, \code{\link{Traj3DStepLengths}}, \code{\link{TrajLength}}
}
