% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{Traj3DRediscretize}
\alias{Traj3DRediscretize}
\title{Resample a 3D trajectory to a constant step length}
\usage{
Traj3DRediscretize(trj3d, R)
}
\arguments{
\item{trj3d}{The 3-dimensional trajectory to be resampled.}

\item{R}{rediscretization step length, in the spatial units of \code{trj}.}
}
\value{
A new 3-dimensional trajectory with a constant segment length which
  follows \code{trj3d}.
}
\description{
Constructs a new 3-dimensional trajectory by resampling the input
3-dimensional trajectory to a fixed step (or segment) length. Timing of frames
is lost, so speed and acceleration cannot be calculated on a rediscretized
trajectory.
}
\details{
Unfortunately this operation is slow for large trajectories.

Based on the appendix in Bovet and Benhamou, (1988), extended to 3
dimensions.
}
\references{
Bovet, P., & Benhamou, S. (1988). Spatial analysis of animals'
  movements using a correlated random walk model. Journal of Theoretical
  Biology, 131(4), 419-433. doi:10.1016/S0022-5193(88)80038-9
}
\seealso{
\code{\link{TrajRediscretize}}
}
