% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{Traj3DSpeed}
\alias{Traj3DSpeed}
\title{Speed along a 3-dimensional trajectory}
\usage{
Traj3DSpeed(vel)
}
\arguments{
\item{vel}{Velocity of a 3-dimensional trajectory, as returned by
\code{\link{Traj3DVelocity}}.}
}
\value{
Numeric vector with speed at each point along the trajectory.
}
\description{
Speed is calculated as the magnitude of velocity. The returned speed will
contain leading and/or trailing `NA` values, depending on the type of
differences used to calculate velocity.
}
\details{
If the trajectory has constant time steps, then the average speed of the
trajectory is `mean(Traj3DSpeed(Traj3DVelocity(trj)), na.rm = TRUE)`. A
trajectory can be resampled by \code{\link{Traj3DResampleTime}} so that it
has constant time steps.

This function is implemented to simply interpret each row of a matrix as a 3D
vector, and return their lengths. Accordingly, this function can also be used
to calculate the magnitude of acceleration as returned by
\code{\link{Traj3DAcceleration}}.
}
\examples{
\dontrun{
# Get the velocity along the trajectory
speed <- Traj3DSpeed(Traj3DVelocity(trj3d))

# Alternative method to calculate mean speed of a portion of a trajectory
si <- 10
ei <- 400
# Mean speed is displacement divided by time
speedMean <- Traj3DLength(trj3d, startIndex = si, endIndex = ei) /
                TrajDuration(trj3d, startIndex = si, endIndex = ei)
}

}
\seealso{
\code{\link{Traj3DVelocity}} to calculate the 3D velocity,
  \code{\link{Traj3DAcceleration}} to calculate the 3D acceleration,
  \code{\link{Traj3DLength}}, \code{\link{TrajDuration}},
  \code{\link{Traj3DResampleTime}}
}
