% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{Traj3DStepLengths}
\alias{Traj3DStepLengths}
\title{3D trajectory step lengths}
\usage{
Traj3DStepLengths(trj3d, startIndex = 1, endIndex = nrow(trj3d))
}
\arguments{
\item{trj3d}{Trajectory to query.}

\item{startIndex}{Index of the starting point.}

\item{endIndex}{Index of the ending point.}
}
\value{
Vector of step lengths. The vector will have length \code{1 - TrajGetNCoords(trj3d)}.
}
\description{
Returns the lengths of each step in a 3-dimensional trajectory (or part of a trajectory).
}
\seealso{
\code{\link{Traj3DFromCoords}}, \code{\link{Traj3DLength}}, \code{\link{TrajStepLengths}}
}
