% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{Traj3DVelocity}
\alias{Traj3DVelocity}
\title{Velocity of a trajectory}
\usage{
Traj3DVelocity(trj3d, diff = c("central", "forward", "backward"))
}
\arguments{
\item{trj3d}{3-dimensional trajectory whose velocity is to be calculated.}

\item{diff}{Type of difference to be calculated, one of "central" (the
default), "forward" or "backward".}
}
\value{
A numeric matrix. Each row represents the 3D velocity vector at each
  point along the trajectory, with `x`, `y` and `z` columns. If \code{diff}
  is \code{"central"}, the first and last velocity values will be \code{NA}
  since velocity cannot be calculated for them. If \code{diff} is
  \code{"forward"}, the last value will be NA, and if \code{diff} is
  \code{"backward"}, the first value will be NA.
}
\description{
The velocity, as a 3-dimensional vector, is approximated at each point of the
trajectory using first-order finite differences. Central, forward or backward
differences can be used. Central differences yield a more accurate
approximation if the velocity is smooth. As a practical guide, if velocity
doesn't change much between steps, use central differences. If it changes
substantially (and not just as an artifact of recording noise), then use
either forward or backward differences.
}
\details{
Intuitively, think of the central difference velocity at a point as the mean
of the velocities of the two adjacent steps. Forward difference velocity is
the velocity of the step starting at the point. Backward difference is the
velocity of the step ending at the point.

Speed (i.e. the magnitude of the velocity) can be derived from velocity by
calling \code{\link{Traj3DSpeed}}.
}
\seealso{
\code{\link{Traj3DSpeed}} for calculating scalar speed, which is the
  magnitude of the velocity vector at each step;
  \code{\link{Traj3DResampleTime}} and \code{\link{Traj3DRediscretize}} to
  resample a trajectory to fixed time or length steps; Finite differences on
  \href{https://en.wikipedia.org/wiki/Finite_difference}{Wikipedia}.
}
