% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajArg}
\alias{TrajArg}
\title{Substitute for the R base function `Arg(z)` that treats the argument of zero
as undefined. The base R function returns 0.}
\usage{
TrajArg(z)
}
\description{
Substitute for the R base function `Arg(z)` that treats the argument of zero
as undefined. The base R function returns 0.
}
