% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajDAFindFirstMaximum}
\alias{TrajDAFindFirstMaximum}
\title{First direction autocorrelation maximum}
\usage{
TrajDAFindFirstMaximum(corr, windowSize = 10)
}
\arguments{
\item{corr}{Direction autocorrelation of a trajectory.}

\item{window}{Size of window used to define what constitutes a local maximum.}
}
\value{
Numeric vector with 2 values, \code{deltaS} and \code{C}, or else
  NULL if there is no local minimum (other than the start point).
}
\description{
Determines the coordinates of the first local maximum for \code{C} in the
direction autocorrelation function of a trajectory as returned by
\code{\link{TrajDirectionAutocorrelations}}. The start point is excluded from
consideration as maximum.
}
\examples{
# Calculate direction autocorrelation for trj
corr <- \\code{\\link{TrajDirectionAutocorrelations}}(trj)
# Extract first local maximum from autocorrelation
minPt <- TrajDAFindFirstMaximum(corr, 50)

# Plot the autocorrelation function
plot(corr)
# Plot a red dot with a black outline at the first maximum
points(minPt["deltaS"], minPt["C"], pch = 16, col = "red", lwd = 2)
points(minPt["deltaS"], minPt["C"], col = "black", lwd = 2)

}
\seealso{
\code{\link{TrajDAFindFirstMinimum}}
}
