% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajDAFindFirstMinimum}
\alias{TrajDAFindFirstMinimum}
\title{First direction autocorrelation minimum}
\usage{
TrajDAFindFirstMinimum(corr, windowSize = 10)
}
\arguments{
\item{corr}{Direction autocorrelation of a trajectory.}

\item{window}{Size of window used to define what constitutes a local mimimum.}
}
\value{
Numeric vector with 2 values, \code{deltaS} and \code{C}, or else
  NULL if there is no local minimum (other than the end point).
}
\description{
Determines the coordinates of the first local minimum for \code{C} in the
direction autocorrelation function of a trajectory as returned by
\code{\link{TrajDirectionAutocorrelations}}. The end point is excluded from
consideration as minimum.
}
\examples{
# Calculate direction autocorrelation for trj
corr <- \\code{\\link{TrajDirectionAutocorrelations}}(trj)
# Extract first local minimum from autocorrelation
minPt <- TrajDAFindFirstMinimum(corr, 50)

# Plot the autocorrelation function
plot(corr)
# Plot a red dot with a black outline at the first minimum
points(minPt["deltaS"], minPt["C"], pch = 16, col = "red", lwd = 2)
points(minPt["deltaS"], minPt["C"], col = "black", lwd = 2)

}
\seealso{
\code{\link{TrajDAFindFirstMaximum}}
}
