% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajDerivatives}
\alias{TrajDerivatives}
\title{Calculates speed and linear acceleration along a track over time.}
\usage{
TrajDerivatives(trj)
}
\value{
list with values \code{speed}, \code{speedTimes}, \code{acceleration}, \code{accelerationTimes}
}
\description{
Calculates speed and linear acceleration along a track over time.
}
