% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajDistance}
\alias{TrajDistance}
\title{Calculates the distance between the start and end of a trajectory.
Also called the diffusion distance, net distance, or bee-line from start to finish.}
\usage{
TrajDistance(trj)
}
\description{
Calculates the distance between the start and end of a trajectory.
Also called the diffusion distance, net distance, or bee-line from start to finish.
}
