% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajDuration}
\alias{TrajDuration}
\title{Trajectory duration}
\usage{
TrajDuration(trj, startIndex = 1, endIndex = nrow(trj))
}
\arguments{
\item{trj}{Trajectory whose duration is to be calculated.}

\item{startIndex}{Index of the starting point.}

\item{endIndex}{Index of the ending point.}
}
\value{
Numeric duration of the trajectory, in time units.
}
\description{
Calculates the temporal duration of a trajectory (or a portion of a
trajectory).
}
\seealso{
\code{\link{TrajGetTimeUnits}}
}
