% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajEmax}
\alias{TrajEmax}
\title{E-max from Cheung et al., (2007)}
\usage{
TrajEmax(trj, eMaxB = FALSE)
}
\arguments{
\item{trj}{Trajectory to be analysed.}

\item{eMaxB}{If TRUE, calculates and returns E-max b, otherwise return E-max.}
}
\value{
E-max for \code{trj}.
}
\description{
Not yet tested
}
