% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{TrajFractalDimensionValues}
\alias{TrajFractalDimensionValues}
\title{Fractal dimension calculation}
\usage{
TrajFractalDimensionValues(trj, stepSizes, adjustD = TRUE)
}
\arguments{
\item{trj}{Trajectory to calculate fractal dimension for.}

\item{stepSizes}{Vector of step sizes used to calculate path lengths (see
\code{\link{TrajFractalStepSizes}}).}

\item{adjustD}{If TRUE, path length is adjusted to reduce truncation error
(Nams, 2006).}
}
\value{
Data frame with columns \code{stepsize} (\eqn{\delta}) and
  \code{pathlength} ((\eqn{L(\delta)}).
}
\description{
Calculates path length (\eqn{L(\delta)}) for a range of step sizes
(\eqn{\delta}). For a fractal curve, \eqn{log(L(\delta))} grows linearly as
\eqn{log(\delta)} grows smaller.
}
\examples{
set.seed(42)
trj <- TrajGenerate()
plot(TrajFractalDimensionValues(trj, TrajFractalStepSizes(trj)), log = "xy", pch = 16, cex = .5)

}
\references{
Nams, V. O. (2006). Improving Accuracy and Precision in Estimating Fractal
  Dimension of Animal movement paths. Acta Biotheoretica, 54(1), 1-11.
  doi:10.1007/s10441-006-5954-8
}
\seealso{
\code{\link{TrajFractalStepSizes}} for a possible calculation of
  \code{stepSizes}. \code{\link{TrajFractalDimension}} for fractal dimension
  calculation.
}
