% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{TrajFractalStepSizes}
\alias{TrajFractalStepSizes}
\title{Fractal dimension step sizes}
\usage{
TrajFractalStepSizes(trj, smallestStepFactor = 1e-04,
  largestStepFactor = 0.1, numSteps = 100)
}
\arguments{
\item{trj}{Trajectory to return step sizes for.}

\item{smallestStepFactor}{Multiplied by the path length of \code{trj} to
determine the smallest step size in the returned range.}

\item{largestStepFactor}{Multiplied by the path length of \code{trj} to
determine the largest step size in the returned range.}

\item{numSteps}{The number of step sizes in the returned vector.}
}
\value{
A vector of step sizes which can be passed to
  \code{\link{TrajFractalDimension}}.
}
\description{
Calculates a vector of possible fractal dimension step sizes, suitable for
use when calculating the fractal dimension of a trajectory. Note that the
selected step sizes may determine the calculated fractal dimension, which may
render fractal dimension meaningless for animal trajectories.
}
\details{
By default, returns a vector of step sizes ranging from 1/1000th of the path
length to 1/10th of the path length, with logarithmically increasing intervals.
}
\seealso{
\code{\link{TrajFractalDimension}}
}
