% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajFromCoords}
\alias{TrajFromCoords}
\title{Create a Trajectory Object}
\usage{
TrajFromCoords(track, xCol = 1, yCol = 2, timeCol = NULL, fps = 50)
}
\arguments{
\item{track}{data frame containing cartesian coordinates and optionally times
for the poiints in the trajectory.}

\item{xCol}{Name or index of the \code{x} column in \code{track} (default 1).}

\item{yCol}{Name or index of the \code{y} column in \code{track} (default 2).}

\item{timeCol}{optional name or index of the column which contains frame
times.}

\item{fps}{Frames per second - used to calculate relative frame times if
\code{track} does not contain a \code{time} column.}
}
\value{
An object with class "\code{Trajectory}", which is a data.frame with
  the following components: \item{x}{X coordinates of trajectory points.}
  \item{y}{Y coordinates of trajectory points.} \item{time}{Time (secs) for
  each point. if \code{timeCol} is specified, values are
  \code{trj[,timeCol]}, otherwise values are calculated from \code{fps}.}
  \item{displacementTime}{Frame times, with frame 1 at time \code{0}.}
  \item{polar}{Coordinates represented as complex number, to simplify working
  with segment angles.} \item{displacement}{Displacements between each pair
  of consecutive points.}
}
\description{
\code{TrajFromCoords} creates a new trajectory object from a set of
2-dimensional cartesian coordinates and some metadata.
}
\details{
If \code{timeCol} is specified, it is assumed to contain the time (in
seconds) of each data point. Otherwise times are calculated for each point as
\code{(frame - 1) / fps} where \code{frame} is the index of the point.
}
