% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajGetFPS}
\alias{TrajGetFPS}
\title{Trajectory frmes-per-second}
\usage{
TrajGetFPS(trj)
}
\arguments{
\item{trj}{Trajectory to query}
}
\description{
Returns the frames-per-second recorded for this trajectory.
}
