% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajGetNCoords}
\alias{TrajGetNCoords}
\title{Trajectory number of coordinates}
\usage{
TrajGetNCoords(trj)
}
\arguments{
\item{trj}{Trajectory to query}
}
\description{
Returns the number of coordinates recorded for this trajectory, i.e. 1 more
than the number of steps.
}
