% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajGetTimeUnits}
\alias{TrajGetTimeUnits}
\title{Trajectory temporal units}
\usage{
TrajGetTimeUnits(trj)
}
\arguments{
\item{trj}{Trajectory to query}
}
\description{
Returns the temporal units specified for a scaled trajectory.
}
\seealso{
\code{\link{TrajFromCoords}}, \code{\link{TrajGetUnits}}.
}
