% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajGetUnits}
\alias{TrajGetUnits}
\title{Trajectory spatial units}
\usage{
TrajGetUnits(trj)
}
\arguments{
\item{trj}{Trajectory to query}
}
\description{
Returns the spatial units specified for a scaled trajectory.
}
\seealso{
\code{\link{TrajScale}}
}
