% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajLength}
\alias{TrajLength}
\title{Calculates the cumulative length of a track, which is the distance travelled}
\usage{
TrajLength(trj)
}
\description{
Calculates the cumulative length of a track, which is the distance travelled
}
