% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{TrajLogSequence}
\alias{TrajLogSequence}
\title{Logarithmically spaced sequence}
\usage{
TrajLogSequence(from, to, length.out)
}
\arguments{
\item{from}{Starting value of the sequence.}

\item{to}{End (maximal) value of the sequence.}

\item{length.out}{Desired length of the sequence (non-negative). Rounded up if fractional.}
}
\description{
Convenience function to return a sequence of points which are regularly
spaced when plotted on a logarithmic axis.
}
\seealso{
\code{\link[base]{seq}}
}
