% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajMeanStepLength}
\alias{TrajMeanStepLength}
\title{Mean trajectory step length}
\usage{
TrajMeanStepLength(trj)
}
\arguments{
\item{trj}{Trajectory to query}
}
\description{
Returns the mean segment length of a trajectory
}
