% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajMeanVelocity}
\alias{TrajMeanVelocity}
\title{Trajectory mean velocity}
\usage{
TrajMeanVelocity(trj, startIndex = 1, endIndex = nrow(trj))
}
\arguments{
\item{trj}{Trajectory whose duration is to be calculated.}

\item{startIndex}{Index of the starting point.}

\item{endIndex}{Index of the ending point.}
}
\value{
Numeric duration of the trajectory, in time units.
}
\description{
Calculates the mean or net velocity of a trajectory (or a portion of a
trajectory). Theisis the velocity from the start point to the end point,
ignoring the path that was taken.
}
\seealso{
\code{\link{TrajGetTimeUnits}}
}
