% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajr-package.R
\docType{package}
\name{trajr-package}
\alias{trajr}
\alias{trajr-package}
\title{trajr: Animal Trajectory Analysis}
\description{
A toolbox to assist with statistical analysis of animal trajectories. It provides simple access to algorithms for calculating and assessing a variety of characteristics such as speed and acceleration, as well as multiple measures of straightness or tortuosity. Some support is provided for 3-dimensional trajectories. McLean & Skowron Volponi (2018) \doi{10.1111/eth.12739}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JimMcL/trajr}
  \item Report bugs at \url{https://github.com/JimMcL/trajr/issues}
}

}
\author{
\strong{Maintainer}: Jim McLean \email{jim_mclean@optusnet.com.au}

}
\keyword{internal}
