% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajr.R
\docType{package}
\name{trajr}
\alias{trajr}
\alias{trajr-package}
\title{trajr: trajectory analysis in R}
\description{
A toolkit for the statistical analysis of 2-dimensional animal trajectories.
}
\details{
Trajr operates on trajectories which are arrays of x and y coordinates. It
can be used to calculate characteristics such as velocity and acceleration,
as well as various measures of straightness or tortuosity.

Trajr does not perform object tracking from videos, it operates on existing
arrays of coordinates.
}
\section{Foo functions}{
 The foo functions ...
}

