% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rediscretization.R
\name{TrajRediscretize}
\alias{TrajRediscretize}
\title{Resample a trajectory to a constant path length}
\usage{
TrajRediscretize(trj, R)
}
\arguments{
\item{trj}{The trajectory to be resampled.}

\item{R}{rediscretization step length.}
}
\value{
a new trajectory which follows \code{trj}
}
\description{
Constructs a new trajectory by resampling the input trajectory to a fixed
segment length.
}
\details{
Based on the appendix in Bovet and Benhamou, (1988).

TODO Handle frame times properly
}
