% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajReverse}
\alias{TrajReverse}
\title{Reverse a trajectory}
\usage{
TrajReverse(trj)
}
\arguments{
\item{trj}{The Trajectory to be reversed.}
}
\value{
A copy of \code{trj} with direction reversed.
}
\description{
Reverses the direction of a trajectory, so that the starting point becomes
the last point and vice versa.
}
