% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajRotate}
\alias{TrajRotate}
\title{Rotates a trajectory so that angle(finish - start) == angle}
\usage{
TrajRotate(trj, angle = 0)
}
\description{
Rotates a trajectory so that angle(finish - start) == angle
}
