% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{TrajsBuild}
\alias{TrajsBuild}
\title{Construct multiple trajectories}
\usage{
TrajsBuild(fileNames, fps, scale = NULL, units = NULL, csvStruct = list(x
  = 1, y = 2, time = NULL), smoothP = 3, smoothN = 41, rootDir = NULL,
  csvReadFn = utils::read.csv)
}
\arguments{
\item{fileNames}{Vector of the names of CSV files containing trajectory
coordinates. All of the files must have the same columns. All file names
must be unique.}

\item{fps}{Vector of frames-per-second values corresponding to the
trajectories in \code{fileNames}.}

\item{scale}{Vector of scale values corresponding to the trajectories in
\code{fileNames}. May be specified as character expressions (e.g. "1 /
1200") rather than numeric values. If NULL, the trajectories will not be
scaled.}

\item{units}{Name of spatial coordinate units after scaling, e.g. "m".}

\item{csvStruct}{A list which identifies the columns in each CSV file which
contain x-, y-, and optionally time-values.}

\item{smoothP}{Filter order to be used for Savitzky-Golay smoothing (see
\code{\link{TrajSmoothSG}})}

\item{smoothN}{Filter length to be used for Savitzky-Golay smoothing (must be
odd, see \code{\link{TrajSmoothSG}})}

\item{rootDir}{Optional name of a top level directory which contains the CSV
files. If \code{rootDir} is not NULL, the CSV files may be located anywhere
within \code{rootDir} or its sub-directories, .}

\item{csvReadFn}{Function used to read the CSV files (see
\code{\link[utils]{read.csv}}, \code{\link[utils]{read.csv2}}).}
}
\value{
A list of trajectories.
}
\description{
Reads multiple trajectories from files, performs some basic sanity checks on
them, and optionally smooths and scales them. Attempts to collect and report
errors for multiple trajectories in a single call.
}
\examples{
\dontrun{
# Names of CSV files containing trajectory coordinates
fileNames <- c('xy001.csv', 'xy003.csv', 'xy004.csv')
# The files are all located under this directory
rootDir <- '.'
# Scale is 1 / pixels per metre
scale <- c('1/1200', '1/1350', '1/1300')
# Files have columns y, x
csvStruct <- list(x = 2, y = 1)
# Apply default smoothing, and the files are formatted as conventional CSV,
# so there's no need to specify csvReadFn
trjs <- TrajsBuild(fileNames, fps = rep(50, length(fileNames)),
                   scale = scale, units = "m",
                   csvStruct = csvStruct, rootDir = rootDir)
}

}
\seealso{
\code{\link[utils]{read.csv}}, \code{\link{TrajFromCoords}},
  \code{\link{TrajScale}}, \code{\link{TrajSmoothSG}}
}
