% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convenience.R
\name{TrajsMergeStats}
\alias{TrajsMergeStats}
\title{Merge trajectory characteristics}
\usage{
TrajsMergeStats(trjs, statsFn)
}
\arguments{
\item{trjs}{List of trajectories to be characterised.}

\item{statsFn}{Function to calculate statistics of interest for a single trajectory.}
}
\description{
Builds a data frame by combining rows of statistical values for multiple trajectories.
}
\section{Note}{

Any NULL valued statistics are converted to NAs.
}

