% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajSmoothSG}
\alias{TrajSmoothSG}
\title{Smooth a trajectory using a Savitzky-Golay filter}
\usage{
TrajSmoothSG(trj, p = 3, n = p + 3 - p\%\%2)
}
\arguments{
\item{trj}{The trajectory to be smoothed.}

\item{p}{polynomial order.}

\item{n}{Filter length (or window size), must be an odd number.}
}
\value{
a new trajectory which is a smoothed version of the input trajectory.
}
\description{
Smooths a trajectory using a Savitzky-Golay smoothing filter.
}
\examples{
trj <- TrajSmoothSG(trj, 3, 101)
}
