% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajStepLengths}
\alias{TrajStepLengths}
\title{Trajectory step lengths}
\usage{
TrajStepLengths(trj)
}
\arguments{
\item{trj}{Trajectory to query}
}
\description{
Returns the lengths of each step in a trajectory.
}
