% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajStraightness}
\alias{TrajStraightness}
\title{Straightness of a Trajectory}
\usage{
TrajStraightness(trj)
}
\arguments{
\item{trj}{Trajectory to calculate straightness of.}
}
\value{
The straightness index of \code{trj}.
}
\description{
Calculates the straightness index of a trajectory, \code{D / L}, where
\code{D} is the beeline distance between the first and last points in the
trajectory,and \code{L} is the path length travelled.
}
