% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajTranslate}
\alias{TrajTranslate}
\title{Translate a trajectory}
\usage{
TrajTranslate(trj, dx, dy)
}
\description{
Translate a trajectory
}
\examples{
# Shift a trajectory so that its origin is (10, 15).
# Note that trajectories created by TrajGenerate always start at (0, 0)
set.seed(42)
trj <- TrajGenerate()
trj <- TrajTranslate(trj, 10, 15)

}
