% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{plot.Trajectory}
\alias{plot.Trajectory}
\title{Plot method for trajectories}
\usage{
\method{plot}{Trajectory}(trj, draw.start.pt = TRUE, add = FALSE,
  type = "l", xlim = range(trj$x), ylim = range(trj$y), asp = 1, ...)
}
\arguments{
\item{trj}{The trajectory to be plotted.}

\item{draw.start.pt}{if TRUE, draws a dot at the start point of the trajectory.}

\item{add}{If TRUE, the trajectory is added to the current plot.}

\item{type, xlim, ylim, asp}{plotting parameters with useful defaults.}
}
\description{
The plot method for Trajectory objects.
}
