% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{plot.Trajectory}
\alias{plot.Trajectory}
\title{Plot method for trajectories}
\usage{
\method{plot}{Trajectory}(x, draw.start.pt = TRUE, add = FALSE,
  type = "l", xlim = range(x$x), ylim = range(x$y), asp = 1, ...)
}
\arguments{
\item{x}{An object of class "Trajectory", the trajectory to be plotted.}

\item{draw.start.pt}{if TRUE, draws a dot at the start point of the trajectory.}

\item{add}{If TRUE, the trajectory is added to the current plot.}

\item{type, xlim, ylim, asp}{plotting parameters with useful defaults.}

\item{...}{Additional arguments are passed to \code{\link{plot}}.}
}
\description{
The plot method for Trajectory objects.
}
\examples{
set.seed(42)
trj <- TrajGenerate(angularErrorSd = 1.3)
plot(trj)

}
\seealso{
\code{\link{TrajFromCoords}}
}
