% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{TrajAcceleration}
\alias{TrajAcceleration}
\title{Acceleration of a trajectory}
\usage{
TrajAcceleration(trj)
}
\arguments{
\item{trj}{Trajectory whose acceleration is to be calculated.}
}
\value{
Vector of complex numbers. The modulus (\code{Mod(a)}) is the
  magnitude of the acceleration at each point, and the argument
  (\code{Arg(a)}) is the direction of the acceleration. The vector has an
  attribute, \code{trj}, with the trajectory as its value.
}
\description{
Returns the acceleration of the trajectory approximated for each point of the
trajectory using the second-order central finite differences. The trajectory
must have equal step times. The function \code{\link{TrajResampleTime}} can
be called to convert a trajectory to fixed step times. The first and last
acceleration values are \code{NA} since acceleration cannot be calculated for
them.
}
\examples{
# A function to plot acceleration as arrows (scaled in length)
AccArrows <- function(acc, scale = .001, ...) {
  trj <- attr(acc, "trj")
  graphics::arrows(trj$x, trj$y, trj$x + Re(acc) * scale, trj$y + Im(acc) * scale, ...)
}

# Generate and plot a random trajectory
set.seed(101)
trj <- TrajGenerate(30)
plot(trj)

# Calculate acceleration
acc <- TrajAcceleration(trj)
# Plot acceleration as red arrows at each point. They need to be scaled down to
# fit in the plot, and the arrowhead lengths need to be shortened to look good
AccArrows(acc, scale = .001, col = "red", length = .1)

}
\seealso{
\code{\link{TrajVelocity}} for calculating velocity,
  \code{\link{TrajResampleTime}} to convert a trajectory to fixed step times.
}
