% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajAngles}
\alias{TrajAngles}
\title{Turning angles of a Trajectory}
\usage{
TrajAngles(trj, lag = 1)
}
\arguments{
\item{trj}{the trajectory whose whose angles are to be calculated.}

\item{lag}{Angles between every lag'th segment is calculated.}
}
\description{
Calculates the angles (in radians) of each segment relative to the previous segment.
}
