% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajDirectionalChange}
\alias{TrajDirectionalChange}
\title{Directional change (DC) for an entire track}
\usage{
TrajDirectionalChange(trj, nFrames = 1)
}
\arguments{
\item{trj}{Track to calculate DC for}

\item{nFrames}{frame delta to process - if 1, every frame is process, if 2, every 2nd frame is processed...
Kitamura, T., & Imafuku, M. (2015). Behavioural mimicry in flight path of Batesian intraspecific polymorphic butterfly Papilio polytes. Proceedings of the Royal Society B: Biological Sciences, 282(1809). doi:10.1098/rspb.2015.0483}
}
\description{
Directional change (DC) for an entire track
}
