% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajDirectionAutocorrelations}
\alias{TrajDirectionAutocorrelations}
\title{Direction autocorrelation}
\usage{
TrajDirectionAutocorrelations(trj, deltaSMax = round(nrow(trj)/2))
}
\arguments{
\item{trj}{The trajectory to calculate the directional autocorrelations for.}

\item{deltaSMax}{Maximum delta s to calculate, default is 1/2 the number of
segments in the trajectory.}
}
\value{
a data frame with 2 columns, \code{deltaS} and \code{C}.
}
\description{
Calculates the autocorrelation of the track for the given delta s, based on
(Shamble et al., 2017). \code{trj} must have been rediscretized so that all
segments have the same length. deltaS is specified in number of segments.
}
\details{
Calculates autocorrelations for delta s values ranging from 1 to
length(track)/2. Track should already be rediscretized.
}
