% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajDistance}
\alias{TrajDistance}
\title{Trajectory distance}
\usage{
TrajDistance(trj, startIndex = 1, endIndex = nrow(trj))
}
\arguments{
\item{trj}{Trajectory whose distance is to be calculated.}

\item{startIndex}{Index of the starting point.}

\item{endIndex}{Index of the ending point.}
}
\value{
Numeric distance from the start to the end of the trajectory.
}
\description{
Calculates the distance between the start and end of a trajectory.
Also called the diffusion distance, net distance, or bee-line from start to finish.
}
