% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{TrajFractalDimension}
\alias{TrajFractalDimension}
\title{Calculate fractal dimension of a trajectory}
\usage{
TrajFractalDimension(trj, stepSizes)
}
\arguments{
\item{trj}{Trajectory to calculate fractal dimension for.}

\item{stepSizes}{Vector of step sizes used to calculate path lengths (see \code{\link{TrajFractalStepSizes}}).}
}
\value{
The fractal dimension of the trajectory for the given step sizes.
}
\description{
Calculates the fractal dimension of a trajectory using the 'dividers' method
(Sugihara & May, 1990). This value may be meaningless for animal trajectories
as they may not be fractal curves - see Benhamou (2004).
}
\references{
Sugihara, G., & M. May, R. (1990). Applications of fractals in ecology. Trends in Ecology & Evolution, 5(3), 79-86. doi:10.1016/0169-5347(90)90235-6

Benhamou, S. (2004). How to reliably estimate the tortuosity of an animal's path. Journal of Theoretical Biology, 229(2), 209-220. doi:10.1016/j.jtbi.2004.03.016
}
\seealso{
\code{\link{TrajFractalStepSizes}}, \code{\link{TrajFractalDimensionValues}}.
}
