% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{TrajFractalDimensionValues}
\alias{TrajFractalDimensionValues}
\title{Fractal dimension calculation}
\usage{
TrajFractalDimensionValues(trj, stepSizes)
}
\arguments{
\item{trj}{Trajectory to calculate fractal dimension for.}

\item{stepSizes}{Vector of step sizes used to calculate path lengths (see \code{\link{TrajFractalStepSizes}}).}
}
\value{
Data frame with columns \code{stepsize} (\eqn{\delta}) and \code{pathlength} ((\eqn{L(\delta)}).
}
\description{
Calculates path length (\eqn{L(\delta)}) for a range of step sizes (\eqn{\delta}).
}
\examples{
set.seed(42)
trj <- TrajGenerate()
plot(TrajFractalDimensionValues(trj, TrajFractalStepSizes(trj)), log = "xy", pch = 16, cex = .5)

}
\seealso{
\code{\link{TrajFractalStepSizes}} for a possible calculation of \code{stepSizes}.
\code{\link{TrajFractalDimension}} for fractal dimension calculation.
}
