% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractal.R
\name{TrajFractalStepSizes}
\alias{TrajFractalStepSizes}
\title{Calculate possible fractal dimension step sizes}
\usage{
TrajFractalStepSizes(trj, smallestStepFactor = 1e-04,
  largestStepFactor = 0.1, numSteps = 100)
}
\description{
Calculate possible fractal dimension step sizes
}
