% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajFromCoords}
\alias{TrajFromCoords}
\title{Create a Trajectory Object}
\usage{
TrajFromCoords(trj, xCol = 1, yCol = 2, timeCol = NULL, fps = 50)
}
\arguments{
\item{trj}{data frame containing x & y columns.}

\item{xCol}{Name or index of the \code{x} column in \code{track} (default 1).}

\item{yCol}{Name or index of the \code{y} column in \code{track} (default 2).}

\item{fps}{Frames per second - used to calculate relative frame times if
\code{track} does not contain a \code{time} column.}
}
\value{
A new trajectory object.
}
\description{
\code{TrajFromCoords} creates a new trajectory object from a set of
2-dimensional cartesian coordinates and some metadata.
}
\details{
If \code{track} contains a \code{time} (or \code{Time}) column, it is assumed
to contain the time (in seconds) of each data point. Otherwise times are
calculated for each point as \code{(frame - 1) / fps} where \code{frame} is
the index of the point.
}
