% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{TrajGenerate}
\alias{TrajGenerate}
\title{Generate a random trajectory}
\usage{
TrajGenerate(n = 1000, random = FALSE, stepLength = 2,
  angularErrorSd = 0.5, linearErrorSd = 0.2, fps = 50)
}
\arguments{
\item{n}{Number of steps in the trajectory.}

\item{random}{If TRUE, a random search trajectory is returned, otherwise a
directory trajectory is returned.}

\item{stepLength}{Mean length of each step in the trajectory, in arbitrary
length units.}

\item{angularErrorSd}{Standard deviation of angular errors in radians.}

\item{linearErrorSd}{Standard deviation of linear step length errors.}

\item{fps}{Simulated frames-per-second - used to generate times for each
point in the trajectory.}
}
\value{
A new Trajectory with \code{n} segments and \code{n + 1} coordinate pairs.
}
\description{
Generates a trajectory. If \code{random} is \code{TRUE}, the trajectory will
be a random walk/idiothetic directed walk, corresponding to an animal
navigating without a compass. If \code{random} is \code{TRUE}, it will be a
directed walk/allothetic directed walk/oriented path, corresponding to an
animal navigating with a compass (Cheung, Zhang, Stricker, & Srinivasan,
2007, 2008).
}
\details{
For both random and directed walks, errors are normally distributed,
unbiased, and independent of each other, so are \emph{simple} walks in the
terminology of Cheung, Zhang, Stricker, & Srinivasan, (2008).
}
\references{
Cheung, A., Zhang, S., Stricker, C., & Srinivasan, M. V. (2007). Animal
navigation: the difficulty of moving in a straight line. Biological
Cybernetics, 97(1), 47-61. doi:10.1007/s00422-007-0158-0

Cheung, A., Zhang, S., Stricker, C., & Srinivasan, M. V. (2008). Animal
navigation: general properties of directed walks. Biological Cybernetics,
99(3), 197-217. doi:10.1007/s00422-008-0251-z
}
