% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate.R
\name{TrajGenerate}
\alias{TrajGenerate}
\title{Generate a random trajectory}
\usage{
TrajGenerate(n = 1000, random = FALSE, stepLength = 2,
  angularErrorSd = 0.5, linearErrorSd = 0.2, fps = 50)
}
\arguments{
\item{n}{Number of steps in the trajectory.}

\item{random}{If TRUE, a random search trajectory is returned, otherwise a directory trajectory is returned.}

\item{stepLength}{Mean length of each step in the trajectory, in arbitrary length units.}

\item{angularErrorSd}{Standard deviation of angular errors in radians.}

\item{linearErrorSd}{Standard deviation of linear step length errors.}

\item{fps}{Simulated frames-per-second - used to generate times for each coordinate in the trajectory.}
}
\value{
A new Trajectory.
}
\description{
Generates a trajectory - either a random search or a directed path.
}
