% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajectory.R
\name{TrajGetNFrames}
\alias{TrajGetNFrames}
\title{Returns the number of frames recorded for this trajectory}
\usage{
TrajGetNFrames(trj)
}
\description{
Returns the number of frames recorded for this trajectory
}
