% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{TrajLength}
\alias{TrajLength}
\title{Path length}
\usage{
TrajLength(trj)
}
\arguments{
\item{trj}{Trajectory whose length is to be calculated.}
}
\value{
Path Numeric length of the trajectory.
}
\description{
Calculates the cumulative length of a track, which is the total distance
travelled along the trajectory.
}
