% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajMeanVectorOfTurningAngles}
\alias{TrajMeanVectorOfTurningAngles}
\title{Mean vector of turning angles}
\usage{
TrajMeanVectorOfTurningAngles(trj, compass.direction = NULL)
}
\arguments{
\item{trj}{Trajectory object.}

\item{compass.direction}{If not \code{NULL}, step angles are calculated
relative to this angle (in radians), otherwise they are calculated relative
to the previous step angle.}
}
\value{
complex number \code{r} which represents the mean vector, Mod(r) is
  the length of the mean vector, Arg(r) is the angle.
}
\description{
Returns the mean vector of the turning angles, as defined by Batschelet,
(1981). A unit vector is created for each turning angle in the trajectory,
and the centre-of-mass/mean vector is returned.
}
