% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajSinuosity}
\alias{TrajSinuosity}
\title{Sinuosity of a trajectory}
\usage{
TrajSinuosity(trj)
}
\arguments{
\item{trj}{Trajectory to calculate sinuosity of.}
}
\value{
The sinuosity of \code{trj}.
}
\description{
Calculates the sinuosity of a trajectory as defined by Benhamou (2004), which
is: \deqn{S = 1.18\sigma_{q} / \sqrtq}.
}
