% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/straightness.R
\name{TrajStraightness}
\alias{TrajStraightness}
\title{Straightness of a Trajectory}
\usage{
TrajStraightness(trj)
}
\arguments{
\item{trj}{Trajectory to calculate straightness of.}
}
\value{
The straightness index of \code{trj}, which is a value between 0
  (infinitely tortuous) to 1 (a straight line).
}
\description{
Calculates the straightness index of a trajectory, \eqn{D / L}, where
\code{D} is the beeline distance between the first and last points in the
trajectory,and \code{L} is the path length travelled.
}
\references{
Batschelet, E. (1981). Circular statistics in biology. ACADEMIC
PRESS, 111 FIFTH AVE., NEW YORK, NY 10003, 1981, 388.

Benhamou, S. (2004). How to reliably estimate the tortuosity of an animal's
path. Journal of Theoretical Biology, 229(2), 209-220.
doi:10.1016/j.jtbi.2004.03.016
}
\seealso{
\code{\link{TrajDistance}} for trajectory distance, and
  \code{\link{TrajLength}} for trajectory path length.
}
