% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speed.R
\name{TrajVelocity}
\alias{TrajVelocity}
\title{Velocity of a trajectory}
\usage{
TrajVelocity(trj)
}
\arguments{
\item{trj}{Trajectory whose velocity is to be calculated.}
}
\value{
Vector of complex numbers. The modulus (\code{Mod(v)}) is the
  magnitude of the velocity at each point, i.e. the speed, and the argument
  (\code{Arg(v)}) is the direction of the velocity. The vector has an
  attribute, \code{trj}, with the trajectory as its value.
}
\description{
The velocity is approximated at each point of the trajectory using
first-order central finite differences. The trajectory must have equal step
times. The function \code{\link{TrajResampleTime}} can be called to convert a
trajectory to fixed step times. The first and last velocity values are
\code{NA} since velocity cannot be calculated for them.
}
\seealso{
\code{\link{TrajAcceleration}} for calculating acceleration,
  \code{\link{TrajResampleTime}} to convert a trajectory to fixed step times.
}
